Citation -> Csl::Citation
