+++ app/models/publication.rb
@@ -226,7 +226,7 @@ class Publication < ActiveRecord::Base
   end
 
   def update_formatted_citations
-    cite = Citation.new(pub_hash)
+    cite = Csl::Citation.new(pub_hash)
     pub_hash[:apa_citation] = cite.to_apa_citation
     pub_hash[:mla_citation] = cite.to_mla_citation
     pub_hash[:chicago_citation] = cite.to_chicago_citation
