+++ lib/pubmed_harvester.rb
@@ -38,7 +38,7 @@ class PubmedHarvester
     # Turn everything into a pub_hash, then generate citations if needed
     result.map! do |pub|
       pub_hash = (pub.pub_hash if pub.respond_to? :pub_hash) || pub
-      cite = Citation.new(pub_hash)
+      cite = Csl::Citation.new(pub_hash)
 
       pub_hash[:apa_citation] = cite.to_apa_citation unless pub_hash[:apa_citation]
       pub_hash[:mla_citation] = cite.to_mla_citation unless pub_hash[:mla_citation]
