+++ lib/science_wire_client.rb
@@ -289,11 +289,9 @@ class ScienceWireClient
 
     def send_query_and_return_pub_hashes(xml_query)
       xml_results = query_sciencewire(xml_query)
-
       xml_results.xpath('//PublicationItem').map do |sw_xml_doc|
         pub_hash = SciencewireSourceRecord.convert_sw_publication_doc_to_hash(sw_xml_doc)
-        cite = Citation.new(pub_hash)
-
+        cite = Csl::Citation.new(pub_hash)
         pub_hash[:apa_citation] = cite.to_apa_citation
         pub_hash[:mla_citation] = cite.to_mla_citation
         pub_hash[:chicago_citation] = cite.to_chicago_citation
